refactor: move jsr and npm api to a new package
