+++ packages/package-registry/README.md
@@ -1,0 +1,3 @@
+# @kksh/package-registry
+
+This package contains helper functions for interacting with js package registries.
