+++ packages/package-registry/src/npm/models.ts
